+++ src/lcd/base/esp_lcd_jd9365.c
@@ -7,6 +7,7 @@
 #include "soc/soc_caps.h"
 
 #if SOC_MIPI_DSI_SUPPORTED
+#include "ESP_PanelLog.h"
 #include "esp_check.h"
 #include "esp_log.h"
 #include "esp_lcd_panel_commands.h"
@@ -19,20 +20,21 @@
 #include "driver/gpio.h"
 #include "esp_lcd_jd9365.h"
 
-#define JD9365_CMD_PAGE (0xE0)
-#define JD9365_PAGE_USER (0x00)
+#include "esp_lcd_vendor_types.h"
 
-#define JD9365_CMD_DSI_INT0 (0x80)
-#define JD9365_DSI_1_LANE (0x00)
-#define JD9365_DSI_2_LANE (0x01)
-#define JD9365_DSI_3_LANE (0x10)
-#define JD9365_DSI_4_LANE (0x11)
+#define JD9365_CMD_PAGE         (0xE0)
+#define JD9365_PAGE_USER        (0x00)
 
-#define JD9365_CMD_GS_BIT (1 << 0)
-#define JD9365_CMD_SS_BIT (1 << 1)
+#define JD9365_CMD_DSI_INT0     (0x80)
+#define JD9365_DSI_1_LANE       (0x00)
+#define JD9365_DSI_2_LANE       (0x01)
+#define JD9365_DSI_3_LANE       (0x10)
+#define JD9365_DSI_4_LANE       (0x11)
 
-typedef struct
-{
+#define JD9365_CMD_GS_BIT       (1 << 0)
+#define JD9365_CMD_SS_BIT       (1 << 1)
+
+typedef struct {
     esp_lcd_panel_io_handle_t io;
     int reset_gpio_num;
     uint8_t madctl_val; // save current value of LCD_CMD_MADCTL register
@@ -40,9 +42,8 @@ typedef struct
     const esp_lcd_panel_vendor_init_cmd_t *init_cmds;
     uint16_t init_cmds_size;
     uint8_t lane_num;
-    struct
-    {
-        unsigned int reset_level : 1;
+    struct {
+        unsigned int reset_level: 1;
     } flags;
     // To save the original functions of MIPI DPI panel
     esp_err_t (*del)(esp_lcd_panel_t *panel);
@@ -56,17 +57,17 @@ static esp_err_t panel_jd9365_init(esp_lcd_panel_t *panel);
 static esp_err_t panel_jd9365_reset(esp_lcd_panel_t *panel);
 static esp_err_t panel_jd9365_invert_color(esp_lcd_panel_t *panel, bool invert_color_data);
 static esp_err_t panel_jd9365_mirror(esp_lcd_panel_t *panel, bool mirror_x, bool mirror_y);
-static esp_err_t panel_jd9365_swap_xy(esp_lcd_panel_t *panel, bool swap_axes);
-static esp_err_t panel_jd9365_set_gap(esp_lcd_panel_t *panel, int x_gap, int y_gap);
 static esp_err_t panel_jd9365_disp_on_off(esp_lcd_panel_t *panel, bool on_off);
 
 esp_err_t esp_lcd_new_panel_jd9365(const esp_lcd_panel_io_handle_t io, const esp_lcd_panel_dev_config_t *panel_dev_config,
                                    esp_lcd_panel_handle_t *ret_panel)
 {
+    ESP_PANEL_ENABLE_TAG_DEBUG_LOG();
+
     ESP_LOGI(TAG, "version: %d.%d.%d", ESP_LCD_JD9365_VER_MAJOR, ESP_LCD_JD9365_VER_MINOR,
              ESP_LCD_JD9365_VER_PATCH);
     ESP_RETURN_ON_FALSE(io && panel_dev_config && ret_panel, ESP_ERR_INVALID_ARG, TAG, "invalid arguments");
-    jd9365_vendor_config_t *vendor_config = (jd9365_vendor_config_t *)panel_dev_config->vendor_config;
+    esp_lcd_panel_vendor_config_t *vendor_config = (esp_lcd_panel_vendor_config_t *)panel_dev_config->vendor_config;
     ESP_RETURN_ON_FALSE(vendor_config && vendor_config->mipi_config.dpi_config && vendor_config->mipi_config.dsi_bus, ESP_ERR_INVALID_ARG, TAG,
                         "invalid vendor config");
 
@@ -74,8 +75,7 @@ esp_err_t esp_lcd_new_panel_jd9365(const esp_lcd_panel_io_handle_t io, const esp
     jd9365_panel_t *jd9365 = (jd9365_panel_t *)calloc(1, sizeof(jd9365_panel_t));
     ESP_RETURN_ON_FALSE(jd9365, ESP_ERR_NO_MEM, TAG, "no mem for jd9365 panel");
 
-    if (panel_dev_config->reset_gpio_num >= 0)
-    {
+    if (panel_dev_config->reset_gpio_num >= 0) {
         gpio_config_t io_conf = {
             .mode = GPIO_MODE_OUTPUT,
             .pin_bit_mask = 1ULL << panel_dev_config->reset_gpio_num,
@@ -83,8 +83,7 @@ esp_err_t esp_lcd_new_panel_jd9365(const esp_lcd_panel_io_handle_t io, const esp
         ESP_GOTO_ON_ERROR(gpio_config(&io_conf), err, TAG, "configure GPIO for RST line failed");
     }
 
-    switch (panel_dev_config->color_space)
-    {
+    switch (panel_dev_config->color_space) {
     case LCD_RGB_ELEMENT_ORDER_RGB:
         jd9365->madctl_val = 0;
         break;
@@ -96,8 +95,7 @@ esp_err_t esp_lcd_new_panel_jd9365(const esp_lcd_panel_io_handle_t io, const esp
         break;
     }
 
-    switch (panel_dev_config->bits_per_pixel)
-    {
+    switch (panel_dev_config->bits_per_pixel) {
     case 16: // RGB565
         jd9365->colmod_val = 0x55;
         break;
@@ -111,10 +109,6 @@ esp_err_t esp_lcd_new_panel_jd9365(const esp_lcd_panel_io_handle_t io, const esp
         ESP_GOTO_ON_FALSE(false, ESP_ERR_NOT_SUPPORTED, err, TAG, "unsupported pixel width");
         break;
     }
-
-    uint8_t ID[3];
-    ESP_GOTO_ON_ERROR(esp_lcd_panel_io_rx_param(io, 0x04, ID, 3), err, TAG, "read ID failed");
-    ESP_LOGI(TAG, "LCD ID: %02X %02X %02X", ID[0], ID[1], ID[2]);
 
     jd9365->io = io;
     jd9365->init_cmds = vendor_config->init_cmds;
@@ -137,8 +131,6 @@ esp_err_t esp_lcd_new_panel_jd9365(const esp_lcd_panel_io_handle_t io, const esp
     panel_handle->init = panel_jd9365_init;
     panel_handle->reset = panel_jd9365_reset;
     panel_handle->mirror = panel_jd9365_mirror;
-    panel_handle->swap_xy = panel_jd9365_swap_xy;
-    panel_handle->set_gap = panel_jd9365_set_gap;
     panel_handle->invert_color = panel_jd9365_invert_color;
     panel_handle->disp_on_off = panel_jd9365_disp_on_off;
     panel_handle->user_data = jd9365;
@@ -148,10 +140,8 @@ esp_err_t esp_lcd_new_panel_jd9365(const esp_lcd_panel_io_handle_t io, const esp
     return ESP_OK;
 
 err:
-    if (jd9365)
-    {
-        if (panel_dev_config->reset_gpio_num >= 0)
-        {
+    if (jd9365) {
+        if (panel_dev_config->reset_gpio_num >= 0) {
             gpio_reset_pin(panel_dev_config->reset_gpio_num);
         }
         free(jd9365);
@@ -374,8 +364,8 @@ static const esp_lcd_panel_vendor_init_cmd_t vendor_specific_init_default[] = {
     {0x09, (uint8_t[]){0x61}, 1, 0},
     {0x0E, (uint8_t[]){0x48}, 1, 0},
 
-    {0x37, (uint8_t[]){0x58}, 1, 0}, // 全志
-    {0x2B, (uint8_t[]){0x0F}, 1, 0}, // 全志
+    {0x37, (uint8_t[]){0x58}, 1, 0},
+    {0x2B, (uint8_t[]){0x0F}, 1, 0},
 
     {0xE0, (uint8_t[]){0x00}, 1, 0},
 
@@ -391,14 +381,13 @@ static esp_err_t panel_jd9365_del(esp_lcd_panel_t *panel)
 {
     jd9365_panel_t *jd9365 = (jd9365_panel_t *)panel->user_data;
 
-    if (jd9365->reset_gpio_num >= 0)
-    {
+    if (jd9365->reset_gpio_num >= 0) {
         gpio_reset_pin(jd9365->reset_gpio_num);
     }
     // Delete MIPI DPI panel
     jd9365->del(panel);
-    free(jd9365);
     ESP_LOGD(TAG, "del jd9365 panel @%p", jd9365);
+    free(jd9365);
 
     return ESP_OK;
 }
@@ -413,8 +402,10 @@ static esp_err_t panel_jd9365_init(esp_lcd_panel_t *panel)
     bool is_user_set = true;
     bool is_cmd_overwritten = false;
 
-    switch (jd9365->lane_num)
-    {
+    switch (jd9365->lane_num) {
+    case 0:
+        lane_command = JD9365_DSI_2_LANE;
+        break;
     case 1:
         lane_command = JD9365_DSI_1_LANE;
         break;
@@ -432,43 +423,37 @@ static esp_err_t panel_jd9365_init(esp_lcd_panel_t *panel)
         return ESP_ERR_INVALID_ARG;
     }
 
-    ESP_RETURN_ON_ERROR(esp_lcd_panel_io_tx_param(io, JD9365_CMD_PAGE, (uint8_t[]){JD9365_PAGE_USER}, 1), TAG, "send command failed");
-    ESP_RETURN_ON_ERROR(esp_lcd_panel_io_tx_param(io, LCD_CMD_MADCTL, (uint8_t[]){
-                                                                          jd9365->madctl_val,
-                                                                      },
-                                                  1),
-                        TAG, "send command failed");
-    ESP_RETURN_ON_ERROR(esp_lcd_panel_io_tx_param(io, LCD_CMD_COLMOD, (uint8_t[]){
-                                                                          jd9365->colmod_val,
-                                                                      },
-                                                  1),
-                        TAG, "send command failed");
-    ESP_RETURN_ON_ERROR(esp_lcd_panel_io_tx_param(io, JD9365_CMD_DSI_INT0, (uint8_t[]){
-                                                                               lane_command,
-                                                                           },
-                                                  1),
-                        TAG, "send command failed");
+    uint8_t ID[3];
+    ESP_RETURN_ON_ERROR(esp_lcd_panel_io_rx_param(io, 0x04, ID, 3), TAG, "read ID failed");
+    ESP_LOGI(TAG, "LCD ID: %02X %02X %02X", ID[0], ID[1], ID[2]);
+
+    ESP_RETURN_ON_ERROR(esp_lcd_panel_io_tx_param(io, JD9365_CMD_PAGE, (uint8_t[]) {
+        JD9365_PAGE_USER
+    }, 1), TAG, "send command failed");
+    ESP_RETURN_ON_ERROR(esp_lcd_panel_io_tx_param(io, LCD_CMD_MADCTL, (uint8_t[]) {
+        jd9365->madctl_val,
+    }, 1), TAG, "send command failed");
+    ESP_RETURN_ON_ERROR(esp_lcd_panel_io_tx_param(io, LCD_CMD_COLMOD, (uint8_t[]) {
+        jd9365->colmod_val,
+    }, 1), TAG, "send command failed");
+    ESP_RETURN_ON_ERROR(esp_lcd_panel_io_tx_param(io, JD9365_CMD_DSI_INT0, (uint8_t[]) {
+        lane_command,
+    }, 1), TAG, "send command failed");
 
     // vendor specific initialization, it can be different between manufacturers
     // should consult the LCD supplier for initialization sequence code
-    if (jd9365->init_cmds)
-    {
+    if (jd9365->init_cmds) {
         init_cmds = jd9365->init_cmds;
         init_cmds_size = jd9365->init_cmds_size;
-    }
-    else
-    {
+    } else {
         init_cmds = vendor_specific_init_default;
         init_cmds_size = sizeof(vendor_specific_init_default) / sizeof(esp_lcd_panel_vendor_init_cmd_t);
     }
 
-    for (int i = 0; i < init_cmds_size; i++)
-    {
+    for (int i = 0; i < init_cmds_size; i++) {
         // Check if the command has been used or conflicts with the internal
-        if (is_user_set && (init_cmds[i].data_bytes > 0))
-        {
-            switch (init_cmds[i].cmd)
-            {
+        if (is_user_set && (init_cmds[i].data_bytes > 0)) {
+            switch (init_cmds[i].cmd) {
             case LCD_CMD_MADCTL:
                 is_cmd_overwritten = true;
                 jd9365->madctl_val = ((uint8_t *)init_cmds[i].data)[0];
@@ -482,8 +467,7 @@ static esp_err_t panel_jd9365_init(esp_lcd_panel_t *panel)
                 break;
             }
 
-            if (is_cmd_overwritten)
-            {
+            if (is_cmd_overwritten) {
                 is_cmd_overwritten = false;
                 ESP_LOGW(TAG, "The %02Xh command has been used and will be overwritten by external initialization sequence",
                          init_cmds[i].cmd);
@@ -495,8 +479,7 @@ static esp_err_t panel_jd9365_init(esp_lcd_panel_t *panel)
         vTaskDelay(pdMS_TO_TICKS(init_cmds[i].delay_ms));
 
         // Check if the current cmd is the "page set" cmd
-        if ((init_cmds[i].cmd == JD9365_CMD_PAGE) && (init_cmds[i].data_bytes > 0))
-        {
+        if ((init_cmds[i].cmd == JD9365_CMD_PAGE) && (init_cmds[i].data_bytes > 0)) {
             is_user_set = (((uint8_t *)init_cmds[i].data)[0] == JD9365_PAGE_USER);
         }
     }
@@ -513,17 +496,14 @@ static esp_err_t panel_jd9365_reset(esp_lcd_panel_t *panel)
     esp_lcd_panel_io_handle_t io = jd9365->io;
 
     // Perform hardware reset
-    if (jd9365->reset_gpio_num >= 0)
-    {
+    if (jd9365->reset_gpio_num >= 0) {
         gpio_set_level(jd9365->reset_gpio_num, !jd9365->flags.reset_level);
         vTaskDelay(pdMS_TO_TICKS(5));
         gpio_set_level(jd9365->reset_gpio_num, jd9365->flags.reset_level);
         vTaskDelay(pdMS_TO_TICKS(10));
         gpio_set_level(jd9365->reset_gpio_num, !jd9365->flags.reset_level);
         vTaskDelay(pdMS_TO_TICKS(120));
-    }
-    else if (io)
-    { // Perform software reset
+    } else if (io) { // Perform software reset
         ESP_RETURN_ON_ERROR(esp_lcd_panel_io_tx_param(io, LCD_CMD_SWRESET, NULL, 0), TAG, "send command failed");
         vTaskDelay(pdMS_TO_TICKS(120));
     }
@@ -539,12 +519,9 @@ static esp_err_t panel_jd9365_invert_color(esp_lcd_panel_t *panel, bool invert_c
 
     ESP_RETURN_ON_FALSE(io, ESP_ERR_INVALID_STATE, TAG, "invalid panel IO");
 
-    if (invert_color_data)
-    {
+    if (invert_color_data) {
         command = LCD_CMD_INVON;
-    }
-    else
-    {
+    } else {
         command = LCD_CMD_INVOFF;
     }
     ESP_RETURN_ON_ERROR(esp_lcd_panel_io_tx_param(io, command, NULL, 0), TAG, "send command failed");
@@ -561,39 +538,23 @@ static esp_err_t panel_jd9365_mirror(esp_lcd_panel_t *panel, bool mirror_x, bool
     ESP_RETURN_ON_FALSE(io, ESP_ERR_INVALID_STATE, TAG, "invalid panel IO");
 
     // Control mirror through LCD command
-    if (mirror_x)
-    {
+    if (mirror_x) {
         madctl_val |= JD9365_CMD_GS_BIT;
-    }
-    else
-    {
+    } else {
         madctl_val &= ~JD9365_CMD_GS_BIT;
     }
-    if (mirror_y)
-    {
+    if (mirror_y) {
         madctl_val |= JD9365_CMD_SS_BIT;
-    }
-    else
-    {
+    } else {
         madctl_val &= ~JD9365_CMD_SS_BIT;
     }
 
-    ESP_RETURN_ON_ERROR(esp_lcd_panel_io_tx_param(io, LCD_CMD_MADCTL, (uint8_t[]){madctl_val}, 1), TAG, "send command failed");
+    ESP_RETURN_ON_ERROR(esp_lcd_panel_io_tx_param(io, LCD_CMD_MADCTL, (uint8_t []) {
+        madctl_val
+    }, 1), TAG, "send command failed");
     jd9365->madctl_val = madctl_val;
 
     return ESP_OK;
-}
-
-static esp_err_t panel_jd9365_swap_xy(esp_lcd_panel_t *panel, bool swap_axes)
-{
-    ESP_LOGW(TAG, "swap_xy is not supported by this panel");
-    return ESP_ERR_NOT_SUPPORTED;
-}
-
-static esp_err_t panel_jd9365_set_gap(esp_lcd_panel_t *panel, int x_gap, int y_gap)
-{
-    ESP_LOGE(TAG, "set_gap is not supported by this panel");
-    return ESP_ERR_NOT_SUPPORTED;
 }
 
 static esp_err_t panel_jd9365_disp_on_off(esp_lcd_panel_t *panel, bool on_off)
@@ -602,12 +563,9 @@ static esp_err_t panel_jd9365_disp_on_off(esp_lcd_panel_t *panel, bool on_off)
     esp_lcd_panel_io_handle_t io = jd9365->io;
     int command = 0;
 
-    if (on_off)
-    {
+    if (on_off) {
         command = LCD_CMD_DISPON;
-    }
-    else
-    {
+    } else {
         command = LCD_CMD_DISPOFF;
     }
     ESP_RETURN_ON_ERROR(esp_lcd_panel_io_tx_param(io, command, NULL, 0), TAG, "send command failed");